requirements.txt
PyYAML==6.0.2
websockets==14.1
